Add tests for the Cache class (and fix a bug).
Unit test were added for the Cache class.  In addition, a bug that I
introduced by changing an "Integer" to an "int" was fixed thanks to the
unit tests.

I'll probably continue adding unit tests to make sure everything is
working as a I intend.  It's also a good way to ensure that we are on
the same page with how everything should work.

+++ src/simizer/storage/Cache.java
@@ -69,8 +69,9 @@ public class Cache extends StorageElement {
     // actually contained in the cache.  If it is not, it could cause the
     // write() method above to throw a NullPointerException.  (It will try to
     // delete a null resource.)
-    
-    this.cachedResources.remove(resourceId);
+
+    // casting is necessary so that the value is not treated as an index
+    this.cachedResources.remove((Integer)resourceId);
     this.cachedResources.addLast(resourceId);
   }
